requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary>=2.8
djangorestframework==3.12.4
selenium==3.141.0
PyVirtualDisplay==2.2
# ChromeDriver==91.0.4472.114
# Version  (Official Build) (64-bit)



